--- optuna_integration/_lightgbm_tuner/__init__.py
@@ -1,19 +1,0 @@
-from optuna_integration._lightgbm_tuner._train import train
-from optuna_integration._lightgbm_tuner.optimize import _imports
-from optuna_integration._lightgbm_tuner.optimize import LightGBMTuner
-from optuna_integration._lightgbm_tuner.optimize import LightGBMTunerCV
-
-
-if _imports.is_successful():
-    from optuna_integration._lightgbm_tuner.sklearn import LGBMClassifier
-    from optuna_integration._lightgbm_tuner.sklearn import LGBMModel
-    from optuna_integration._lightgbm_tuner.sklearn import LGBMRegressor
-
-__all__ = [
-    "LightGBMTuner",
-    "LightGBMTunerCV",
-    "LGBMClassifier",
-    "LGBMModel",
-    "LGBMRegressor",
-    "train",
-]
