+++ optuna_integration/lightgbm/__init__.py
@@ -2,7 +2,7 @@ import sys
 
 from optuna._imports import try_import
 
-from optuna_integration import _lightgbm_tuner as tuner
+from optuna_integration.lightgbm import _lightgbm_tuner as tuner
 
 from .lightgbm import LightGBMPruningCallback
 
@@ -14,8 +14,8 @@ with try_import() as _imports:
 if _imports.is_successful():
     # To pass tests/lightgbm_tuner_tests/test_optimize.py.
     from lightgbm import Dataset
-    from optuna_integration._lightgbm_tuner import LightGBMTuner
-    from optuna_integration._lightgbm_tuner import LightGBMTunerCV
+    from optuna_integration.lightgbm._lightgbm_tuner import LightGBMTuner
+    from optuna_integration.lightgbm._lightgbm_tuner import LightGBMTunerCV
 
     _names_from_tuners = ["train", "LGBMModel", "LGBMClassifier", "LGBMRegressor"]
 
@@ -27,6 +27,7 @@ if _imports.is_successful():
 
     # API from lightgbm_tuner.
     for api_name in _names_from_tuners:
+        print(tuner.__dict__.keys())
         setattr(sys.modules[__name__], api_name, tuner.__dict__[api_name])
 else:
     # To create docstring of train.
