+++ optuna_integration/lightgbm/_lightgbm_tuner/optimize.py
@@ -26,8 +26,8 @@ from optuna.trial import FrozenTrial
 from optuna.trial import TrialState
 import tqdm
 
-from optuna_integration._lightgbm_tuner.alias import _handling_alias_metrics
-from optuna_integration._lightgbm_tuner.alias import _handling_alias_parameters
+from optuna_integration.lightgbm._lightgbm_tuner.alias import _handling_alias_metrics
+from optuna_integration.lightgbm._lightgbm_tuner.alias import _handling_alias_parameters
 
 
 with try_import() as _imports:
@@ -60,9 +60,11 @@ _DEFAULT_LIGHTGBM_PARAMETERS = {
 
 _logger = optuna.logging.get_logger(__name__)
 
+
 class _CustomObjectiveType(Protocol):
     def __call__(self, preds: np.ndarray, train: "lgb.Dataset") -> tuple[np.ndarray, np.ndarray]:
         raise NotImplementedError
+
 
 def _get_custom_objective(lgbm_kwargs: dict[str, Any]) -> _CustomObjectiveType | None:
     objective = lgbm_kwargs.get("objective")
