+++ tests/lightgbm/test_optimize.py
@@ -15,12 +15,13 @@ from optuna.study import Study
 import pytest
 
 import optuna_integration
-from optuna_integration._lightgbm_tuner.optimize import _BaseTuner
-from optuna_integration._lightgbm_tuner.optimize import _OptunaObjective
-from optuna_integration._lightgbm_tuner.optimize import _OptunaObjectiveCV
-from optuna_integration._lightgbm_tuner.optimize import LightGBMTuner
-from optuna_integration._lightgbm_tuner.optimize import LightGBMTunerCV
+import optuna_integration.lightgbm
 import optuna_integration.lightgbm as lgb
+from optuna_integration.lightgbm._lightgbm_tuner.optimize import _BaseTuner
+from optuna_integration.lightgbm._lightgbm_tuner.optimize import _OptunaObjective
+from optuna_integration.lightgbm._lightgbm_tuner.optimize import _OptunaObjectiveCV
+from optuna_integration.lightgbm._lightgbm_tuner.optimize import LightGBMTuner
+from optuna_integration.lightgbm._lightgbm_tuner.optimize import LightGBMTunerCV
 
 
 with try_import():
@@ -601,7 +602,7 @@ class TestLightGBMTuner:
         params: dict = {"verbose": -1}
         dataset = lgb.Dataset(np.zeros((10, 10)))
 
-        with mock.patch("optuna_integration._lightgbm_tuner.optimize.os.mkdir") as m:
+        with mock.patch("optuna_integration.lightgbm._lightgbm_tuner.optimize.os.mkdir") as m:
             with mock.patch("os.path.exists", return_value=dir_exists):
                 LightGBMTuner(params, dataset, valid_sets=dataset, model_dir="./booster")
                 assert m.called == expected
@@ -641,7 +642,7 @@ class TestLightGBMTuner:
         def objective(trial: optuna.trial.Trial, value: float) -> float:
             trial.storage.set_trial_system_attr(
                 trial._trial_id,
-                optuna_integration._lightgbm_tuner.optimize._STEP_NAME_KEY,
+                optuna_integration.lightgbm.optimize._STEP_NAME_KEY,
                 "step{:.0f}".format(value),
             )
             return trial.suggest_float("x", value, value)
